+++ cmd/login/auth.go
@@ -24,9 +24,9 @@ type authInfo struct {
 	refreshToken string
 }
 
-// storeCredentials stores the installation's CA certificate, and
+// storeMCCredentials stores the installation's CA certificate, and
 // updates the kubeconfig with the configuration for the k8s api access.
-func storeCredentials(k8sConfigAccess clientcmd.ConfigAccess, i *installation.Installation, authResult authInfo, fs afero.Fs, internalAPI bool) error {
+func storeMCCredentials(k8sConfigAccess clientcmd.ConfigAccess, i *installation.Installation, authResult authInfo, fs afero.Fs, internalAPI bool) error {
 	config, err := k8sConfigAccess.GetStartingConfig()
 	if err != nil {
 		return microerror.Mask(err)
@@ -114,6 +114,85 @@ func storeCredentials(k8sConfigAccess clientcmd.ConfigAccess, i *installation.In
 		return microerror.Mask(err)
 	}
 
+	return nil
+}
+
+// printMCCredentials saves the installation's CA certificate, and
+// writes the configuration for the k8s api access into a separate file.
+func printMCCredentials(k8sConfigAccess clientcmd.ConfigAccess, i *installation.Installation, authResult authInfo, fs afero.Fs, internalAPI bool, filePath string) error {
+	kUsername := fmt.Sprintf("gs-%s-%s", authResult.username, i.Codename)
+	contextName := kubeconfig.GenerateKubeContextName(i.Codename)
+	clusterName := fmt.Sprintf("gs-%s", i.Codename)
+
+	// Store CA certificate.
+	err := kubeconfig.WriteCertificate(i.CACert, clusterName, fs)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	var server string
+	{
+		if internalAPI {
+			server = i.K8sInternalApiURL
+		} else {
+			server = i.K8sApiURL
+		}
+	}
+
+	var certPath string
+	{
+		certPath, err = kubeconfig.GetKubeCertFilePath(clusterName)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+	}
+
+	authInfo := clientcmdapi.NewAuthInfo()
+	{
+		if len(authResult.clientID) > 0 {
+			authInfo.AuthProvider = &clientcmdapi.AuthProviderConfig{
+				Name: "oidc",
+				Config: map[string]string{
+					ClientID:     authResult.clientID,
+					IDToken:      authResult.token,
+					Issuer:       i.AuthURL,
+					RefreshToken: authResult.refreshToken,
+				},
+			}
+		} else {
+			authInfo.Token = authResult.token
+		}
+	}
+
+	kubeconfig := clientcmdapi.Config{
+		APIVersion: "v1",
+		Kind:       "Config",
+		Clusters: map[string]*clientcmdapi.Cluster{
+			contextName: {
+				Server:               server,
+				CertificateAuthority: certPath,
+			},
+		},
+		Contexts: map[string]*clientcmdapi.Context{
+			contextName: {
+				Cluster:  clusterName,
+				AuthInfo: kUsername,
+			},
+		},
+		AuthInfos: map[string]*clientcmdapi.AuthInfo{
+			kUsername: authInfo,
+		},
+		CurrentContext: contextName,
+	}
+	if exists, err := afero.Exists(fs, filePath); exists {
+		return microerror.Maskf(fileExistsError, "The destination file %s already exists. Please specify a different destination.", filePath)
+	} else if err != nil {
+		return microerror.Mask(err)
+	}
+	err = clientcmd.WriteToFile(kubeconfig, filePath)
+	if err != nil {
+		return microerror.Mask(err)
+	}
 	return nil
 }