+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ### Added
 
+- Add support for self-contained kubeconfig creation for management cluster context.
 - Add support for templating App CRs in organization namespace.
 - Add `--catalog-namespace` flag to `template app`.